requirements.txt
googlesearch-python
beautifulsoup4
requests
google-search-results
python-dotenv
fake-useragent
streamlit